@@ -15,50 +15,48 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^7.2.11",
-    "@angular/cdk": "^7.3.6",
-    "@angular/common": "^7.2.11",
-    "@angular/compiler": "^7.2.11",
-    "@angular/core": "^7.2.11",
-    "@angular/forms": "^7.2.11",
-    "@angular/http": "^7.2.11",
-    "@angular/material": "^7.3.6",
-    "@angular/platform-browser": "^7.2.11",
-    "@angular/platform-browser-dynamic": "^7.2.11",
-    "@angular/platform-server": "^7.2.11",
-    "@angular/router": "^7.2.11",
-    "@angular/fire": "^5.1.2",
+    "@angular/animations": "^12.0.2",
+    "@angular/cdk": "^12.0.2",
+    "@angular/common": "^12.0.2",
+    "@angular/compiler": "^12.0.2",
+    "@angular/core": "^12.0.2",
+    "@angular/forms": "^12.0.2",
+    "@angular/material": "^12.0.2",
+    "@angular/platform-browser": "^12.0.2",
+    "@angular/platform-browser-dynamic": "^12.0.2",
+    "@angular/platform-server": "^12.0.2",
+    "@angular/router": "^12.0.2",
+    "@angular/fire": "^6.1.5",
     "@fortawesome/fontawesome-free": "^5.2.0",
     "core-js": "^2.5.0",
-    "firebase": "^5.5.0",
-    "hammerjs": "^2.0.8",
+    "firebase": "^8.6.3",
     "lodash": "^4.17.4",
-    "rxjs": "^6.4.0",
-    "tslib": "^1.9.0",
-    "zone.js": "^0.8.16"
+    "rxjs": "^6.6.7",
+    "tslib": "^2.0.0",
+    "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.13.0",
-    "@angular/cli": "^7.3.7",
-    "@angular/compiler-cli": "^7.2.11",
-    "@types/jasmine": "2.5.38",
+    "@angular-devkit/build-angular": "~12.0.2",
+    "@angular/cli": "^12.0.2",
+    "@angular/compiler-cli": "^12.0.2",
+    "@types/jasmine": "~3.6.0",
     "@types/lodash": "^4.14.57",
-    "@types/node": "~6.0.60",
-    "codelyzer": "^4.5.0",
-    "jasmine-core": "~2.5.2",
-    "jasmine-spec-reporter": "~3.2.0",
-    "karma": "~1.4.1",
-    "karma-chrome-launcher": "~2.0.0",
+    "@types/node": "^12.11.1",
+    "codelyzer": "^6.0.0",
+    "jasmine-core": "~3.6.0",
+    "jasmine-spec-reporter": "~5.0.0",
+    "karma": "~6.3.2",
+    "karma-chrome-launcher": "~3.1.0",
     "karma-cli": "~1.0.1",
-    "karma-coverage-istanbul-reporter": "^0.2.0",
-    "karma-jasmine": "~1.0.2",
-    "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "~5.1.0",
+    "karma-coverage-istanbul-reporter": "~3.0.2",
+    "karma-jasmine": "~4.0.0",
+    "karma-jasmine-html-reporter": "^1.5.0",
+    "protractor": "~7.0.0",
     "ts-node": "~2.0.0",
-    "tslint": "^5.6.0",
+    "tslint": "~6.1.0",
     "tslint-consistent-codestyle": "1.14.1",
     "tslint-eslint-rules": "5.4.0",
     "tslint-sonarts": "^1.6.0",
-    "typescript": "3.2.4"
+    "typescript": "4.2.4"
   }
 }
