@@ -14,7 +14,7 @@ import "core-js/es6/regexp";
 import "core-js/es6/map";
 import "core-js/es6/set";
 import "core-js/es6/reflect";
-import "zone.js/dist/zone";
+import "zone.js";
 
 // If you need to support the browsers/features below, uncomment the import
 // and run `npm install import-name-here';
